+++ .babelrc
@@ -1,0 +1,6 @@
+{
+    "presets": ["@babel/preset-env"],
+    "plugins": [
+        "@babel/plugin-transform-runtime"
+    ]
+}
