+++ package.json
@@ -4,8 +4,8 @@
   "main": "index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "gulp",
-    "build": "gulp build"
+    "start": "cross-env NODE_ENV=development gulp",
+    "build": "cross-env NODE_ENV=production gulp build"
   },
   "repository": {
     "type": "git",
@@ -19,17 +19,25 @@
   "homepage": "https://github.com/zheleznovux/rubius-hw#readme",
   "description": "",
   "devDependencies": {
+    "@babel/core": "^7.13.13",
+    "@babel/plugin-transform-runtime": "^7.13.10",
+    "@babel/preset-env": "^7.13.12",
     "autoprefixer": "^10.2.5",
+    "babel-loader": "^8.2.2",
     "bootstrap": "^4.6.0",
     "browser-sync": "^2.26.14",
+    "cross-env": "^7.0.3",
     "cssnano": "^4.1.10",
     "del": "^6.0.0",
     "gulp": "^4.0.2",
     "gulp-postcss": "^9.0.0",
+    "gulp-rename": "^2.0.0",
     "gulp-sass": "^4.1.0",
     "gulp-sourcemaps": "^3.0.0",
     "jquery": "^3.6.0",
-    "postcss": "^8.2.7"
+    "postcss": "^8.2.7",
+    "webpack": "^5.28.0",
+    "webpack-stream": "^6.1.2"
   },
   "dependencies": {
     "normalize.css": "^8.0.1"
