+++ src/index.js
@@ -1,0 +1,3 @@
+// /fancy
+import './js/main';
+
