+++ src/scss/layout/_prices.scss
@@ -32,6 +32,7 @@
         font-size: 14px;
         cursor: pointer;
         text-align: center;
+        border-bottom: 2px solid transparent;
         padding-bottom: 10px;
 
         &.swiper-slide-thumb-active {
